take out the class from the package object
